+++ app/lib/database/period_database.dart
@@ -1,6 +1,6 @@
 import 'package:sqflite/sqflite.dart';
 import 'package:path/path.dart';
-import 'package:menstrudel/models/period_logs.dart';
+import 'package:menstrudel/models/period_logs/period_logs.dart';
 import 'package:menstrudel/models/periods/period.dart';
 
 class PeriodDatabase {
