+++ app/lib/utils/period_predictor.dart
@@ -1,4 +1,4 @@
-import 'package:menstrudel/models/period_logs.dart';
+import 'package:menstrudel/models/period_logs/period_logs.dart';
 import 'package:menstrudel/models/periods/period.dart';
 import 'package:menstrudel/models/period_prediction_result.dart';
 import 'package:menstrudel/models/cycle_stats.dart';
